src/htmlnode.py
class HtmlNode:
    def __init__(self, tag=None, value=None, children=None, props=None):
        self.tag = tag
        self.value = value
        self.children = children or []
        self.props = props or {}

    def to_html(self):
        raise NotImplementedError("This method must be implemented by a subclass")
    
    def props_to_html(self):
        return " ".join([f"{key}='{value}'" for key, value in self.props.items()])
    
    def __repr__(self):
        return f"HtmlNode({self.tag}, {self.value}, {self.children}, {self.props})"
